First pass on rate-limit/error handling for MB API
Add a JS class to wrap requests to the MusicBrainz API which does rate
limiting of requests with retry and adaptive delay support inspired by
Picard's code. It is initially implemented only for the release lookup
request - I plan to also use it for the login requests and the data
submission.

As part of the implementation, I switched to using JS modules, which
means that some of the variables are no longer present on the window
object and need to be explicitly passed when rendering templates. I
think I got all of them, but I might have missed some, which could
result in some template rendering issues.